Rename ui/elements/metrics.py to dual_value.py
The module contains LargeDualValue and SmallDualValue classes, not
rendering metrics. Rename to dual_value.py so the filename matches the
contents. Update all imports accordingly.
