+++ src/deckboard/presets/sensors.py
@@ -3,6 +3,6 @@
 from __future__ import annotations
 
 from ..ui.cards.status import StatusCard
-from ..ui.elements.metrics import LargeDualValue, SmallDualValue
+from ..ui.elements.dual_value import LargeDualValue, SmallDualValue
 
 __all__ = ["LargeDualValue", "SmallDualValue", "StatusCard"]
